Moved product attribute model out of the product model
